back-end/api_2/requirements.txt
snowflake-connector-python
uvicorn
python-jose
fastapi==0.99.1
pydantic
pymongo